Fix questionnaire chart display issue
This fix addresses certain charts not displaying
when an answer is entered as a string.
The fix creates a nested forloop, but the size
of the inner forloop is limited to how many
unique responses the patient has, so performance
shouldn't be adversely affected.

Jira issue: [ORMS-121]


(cherry picked from commit 2be0a995a81265fcc268c3e21b06e969aa501577)

Co-authored-by: vmatassa <victor.matassa@muhc.mcgill.ca>